feat: restrict typescript interfaces
BREAKING CHANGE: TypeScript will show errors for extraneous API inputs.

Most functions no longer accept an object with any number of additional
key-value pairs. If a key-value pair is provided that does not match the
new restricted interfaces, typescript will alert the developer that the
API is being used incorrectly.

This also includes a few refactors to simplify the code a bit.