feat: simplify documentation interface
The documentation interface is an addon in the storybook essentials
package. Documentation on the package can be found here:
https://www.npmjs.com/package/@storybook/addon-docs.

The API for this is not well defined, it has deeply nested objects, and
there is a different path to document the book versus an individual
story (parameters.docs.description.component vs parameters.docs.description.story).

To make this more user friendly, a user can now add a description in the
same way for a book or a story. By providing an array of objects that
define the link name and value, a "Links" section will be added to the
documentation, and each link will be opened in a new tab when clicked.