feat: simplify decorator interface
BREAKING CHANGE: decorators prop now takes an array of strings

This has the added bonus of fixing a bug where the book level decorator
was included twice at the story level.

Since we pass the results of the book function into the story functions,
storybook will think that we would like the decorator to be applied both
at the book and story level. If we pass in an array of strings, for our
decorators, we can have a more simple interface for our developers, while
also being able to determine which decorators are passed in at the story
level versus the book level.